stream: create class factory
